Added struct neighbor, connect works

+++ Protocol/message.c
@@ -4,7 +4,7 @@
 
 static uint32_t id = 0;
 
-bool create_connect_message(int8_t src_id, int8_t dst_id, message *msg)
+bool create_connect_message(int32_t src_id, int32_t dst_id, message *msg)
 {
     msg->src_id = src_id;
     msg->dst_id = dst_id;
@@ -14,7 +14,7 @@ bool create_connect_message(int8_t src_id, int8_t dst_id, message *msg)
     return true;
 }
 
-bool create_ack_message(int8_t src_id, int8_t dst_id, message *msg)
+bool create_ack_message(int32_t src_id, int32_t dst_id, message *msg)
 {
     msg->src_id = src_id;
     msg->dst_id = dst_id;
